+++ src/main/java/com/nobita/demo/dao/Ingredient.java
@@ -18,7 +18,7 @@ public class Ingredient implements BaseDAO<com.nobita.demo.model.Ingredient>{
     }
 
     @Override
-    public com.nobita.demo.model.Ingredient findByID(int id) {
+    public com.nobita.demo.model.Ingredient findByID(long id) {
         String sql="select i.* ,u.name as name_unit from ingredient i lef join unit u on u.id=i.id_ingredient";
         Object[]values ={id};
         return jdbcTemplate.queryForObject(sql,new IngredientRowMapper(),values);
@@ -39,7 +39,7 @@ public class Ingredient implements BaseDAO<com.nobita.demo.model.Ingredient>{
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(long id) {
         String sql="delete from ingredient where id=?";
         Object[]values ={id};
         return jdbcTemplate.update(sql,values) >0;
