+++ src/main/java/com/nobita/demo/dao/ProductLineDAO.java
@@ -19,7 +19,7 @@ public class ProductLineDAO implements BaseDAO<ProductLine> {
     }
 
     @Override
-    public ProductLine findByID(int id) {
+    public ProductLine findByID(long id) {
         String sql = "select * from productline where id=?";
         Object[] values = {id};
         return jdbcTemplate.queryForObject(sql, new ProductLineRowMapper(), values);
@@ -40,7 +40,7 @@ public class ProductLineDAO implements BaseDAO<ProductLine> {
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(long id) {
         String sql = "delete from productline where id=?";
         Object[] values = {id};
         return jdbcTemplate.update(sql, values) > 0;
