+++ src/main/java/com/nobita/demo/dao/UnitDAO.java
@@ -20,7 +20,7 @@ public class UnitDAO implements BaseDAO<Unit> {
     }
 
     @Override
-    public Unit findByID(int id) {
+    public Unit findByID(long id) {
         String sql = "select * from unit where id=?";
         Object[] values = {id};
         return jdbcTemplate.queryForObject(sql, new UnitRowMapper(), values);
@@ -41,7 +41,7 @@ public class UnitDAO implements BaseDAO<Unit> {
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(long id) {
         String sql = "delete * from unit where id=?";
         Object[] values = {id};
         return jdbcTemplate.update(sql, values) > 0;
